// frontend/src/context/context.types.tsx
import type { User } from "@supabase/supabase-js";

export type AuthContextType = {
	user: User | null;
	loading: boolean;
	login: (email: string, password: string) => Promise<void>;
	signUp: (email: string, password: string) => Promise<void>;
	logout: () => Promise<void>;
	signInWithOAuth: (provider: AuthOAuthProvider) => Promise<void>;
};

export type AuthOAuthProvider = "google" | "discord" | "apple";